e2e reorg helpers
