+++ test/framework/alltypes_helpers.go
@@ -1,5 +1,3 @@
-// +build e2e
-
 /*
 Copyright 2020 The Kubernetes Authors.
 
@@ -16,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package discovery
+package framework
 
 import (
 	"context"
@@ -26,10 +24,9 @@ import (
 	"path"
 	"path/filepath"
 
-	"github.com/onsi/ginkgo"
+	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
-	appsv1 "k8s.io/api/apps/v1"
 	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	apimeta "k8s.io/apimachinery/pkg/api/meta"
@@ -37,34 +34,13 @@ import (
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
 	clusterv1 "sigs.k8s.io/cluster-api/api/v1alpha3"
-	controlplanev1 "sigs.k8s.io/cluster-api/controlplane/kubeadm/api/v1alpha3"
-	"sigs.k8s.io/cluster-api/test/framework"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/yaml"
 )
 
-// Provides methods for discovering Cluster API objects existing in the management cluster.
-
-// GetControllerDeploymentsInput is the input for GetControllerDeployments.
-type GetControllerDeploymentsInput struct {
-	Lister framework.Lister
-}
-
-// GetControllerDeployments returns all the deployment for the cluster API controllers existing in a management cluster.
-func GetControllerDeployments(ctx context.Context, input GetControllerDeploymentsInput) []*appsv1.Deployment {
-	deploymentList := &appsv1.DeploymentList{}
-	Expect(input.Lister.List(ctx, deploymentList, capiProviderOptions()...)).To(Succeed(), "Failed to list deployments for the cluster API controllers")
-
-	deployments := make([]*appsv1.Deployment, len(deploymentList.Items))
-	for i := range deploymentList.Items {
-		deployments[i] = &deploymentList.Items[i]
-	}
-	return deployments
-}
-
 // GetCAPIResourcesInput is the input for GetCAPIResources.
 type GetCAPIResourcesInput struct {
-	Lister    framework.Lister
+	Lister    Lister
 	Namespace string
 }
 
@@ -88,7 +64,7 @@ func GetCAPIResources(ctx context.Context, input GetCAPIResourcesInput) []*unstr
 			if apierrors.IsNotFound(err) {
 				continue
 			}
-			ginkgo.Fail(fmt.Sprintf("failed to list %q resources: %v", typeList.GroupVersionKind(), err))
+			Fail(fmt.Sprintf("failed to list %q resources: %v", typeList.GroupVersionKind(), err))
 		}
 		for i := range typeList.Items {
 			obj := typeList.Items[i]
@@ -101,7 +77,7 @@ func GetCAPIResources(ctx context.Context, input GetCAPIResourcesInput) []*unstr
 
 // getClusterAPITypes returns the list of TypeMeta to be considered for the the move discovery phase.
 // This list includes all the types belonging to CAPI providers.
-func getClusterAPITypes(ctx context.Context, lister framework.Lister) []metav1.TypeMeta {
+func getClusterAPITypes(ctx context.Context, lister Lister) []metav1.TypeMeta {
 	discoveredTypes := []metav1.TypeMeta{}
 
 	crdList := &apiextensionsv1.CustomResourceDefinitionList{}
@@ -126,68 +102,9 @@ func getClusterAPITypes(ctx context.Context, lister framework.Lister) []metav1.T
 	return discoveredTypes
 }
 
-// GetClusterByNameInput is the input for GetClusterByName.
-type GetClusterByNameInput struct {
-	Getter    framework.Getter
-	Name      string
-	Namespace string
-}
-
-// GetClusterByName returns a Cluster object given his name
-func GetClusterByName(ctx context.Context, input GetClusterByNameInput) *clusterv1.Cluster {
-	cluster := &clusterv1.Cluster{}
-	key := client.ObjectKey{
-		Namespace: input.Namespace,
-		Name:      input.Name,
-	}
-	Expect(input.Getter.Get(ctx, key, cluster)).To(Succeed(), "Failed to get Cluster object %s/%s", input.Namespace, input.Name)
-	return cluster
-}
-
-// GetKubeadmControlPlaneByClusterInput is the input for GetKubeadmControlPlaneByCluster.
-type GetKubeadmControlPlaneByClusterInput struct {
-	Lister      framework.Lister
-	ClusterName string
-	Namespace   string
-}
-
-// GetKubeadmControlPlaneByCluster returns the KubeadmControlPlane objects for a cluster.
-// Important! this method relies on labels that are created by the CAPI controllers during the first reconciliation, so
-// it is necessary to ensure this is already happened before calling it.
-func GetKubeadmControlPlaneByCluster(ctx context.Context, input GetKubeadmControlPlaneByClusterInput) *controlplanev1.KubeadmControlPlane {
-	controlPlaneList := &controlplanev1.KubeadmControlPlaneList{}
-	Expect(input.Lister.List(ctx, controlPlaneList, byClusterOptions(input.ClusterName, input.Namespace)...)).To(Succeed(), "Failed to list KubeadmControlPlane object for Cluster %s/%s", input.Namespace, input.ClusterName)
-	Expect(len(controlPlaneList.Items)).ToNot(BeNumerically(">", 1), "Cluster %s/%s should not have more than 1 KubeadmControlPlane object", input.Namespace, input.ClusterName)
-	if len(controlPlaneList.Items) == 1 {
-		return &controlPlaneList.Items[0]
-	}
-	return nil
-}
-
-// GetMachineDeploymentsByClusterInput is the input for GetMachineDeploymentsByCluster.
-type GetMachineDeploymentsByClusterInput struct {
-	Lister      framework.Lister
-	ClusterName string
-	Namespace   string
-}
-
-// GetMachineDeploymentsByCluster returns the MachineDeployments objects for a cluster.
-// Important! this method relies on labels that are created by the CAPI controllers during the first reconciliation, so
-// it is necessary to ensure this is already happened before calling it.
-func GetMachineDeploymentsByCluster(ctx context.Context, input GetMachineDeploymentsByClusterInput) []*clusterv1.MachineDeployment {
-	deploymentList := &clusterv1.MachineDeploymentList{}
-	Expect(input.Lister.List(ctx, deploymentList, byClusterOptions(input.ClusterName, input.Namespace)...)).To(Succeed(), "Failed to list MachineDeployments object for Cluster %s/%s", input.Namespace, input.ClusterName)
-
-	deployments := make([]*clusterv1.MachineDeployment, len(deploymentList.Items))
-	for i := range deploymentList.Items {
-		deployments[i] = &deploymentList.Items[i]
-	}
-	return deployments
-}
-
 // DumpAllResourcesInput is the input for DumpAllResources.
 type DumpAllResourcesInput struct {
-	Lister    framework.Lister
+	Lister    Lister
 	Namespace string
 	LogPath   string
 }
@@ -238,12 +155,20 @@ func capiProviderOptions() []client.ListOption {
 	}
 }
 
-// byClusterOptions returns a set of ListOptions that allows to identify all the objects belonging to a Cluster.
-func byClusterOptions(name, namespace string) []client.ListOption {
-	return []client.ListOption{
-		client.InNamespace(namespace),
-		client.MatchingLabels{
-			clusterv1.ClusterLabelName: name,
-		},
+// CreateRelatedResourcesInput is the input type for CreateRelatedResources.
+type CreateRelatedResourcesInput struct {
+	Creator          Creator
+	RelatedResources []runtime.Object
+}
+
+// CreateRelatedResources is used to create runtime.Objects.
+func CreateRelatedResources(ctx context.Context, input CreateRelatedResourcesInput, intervals ...interface{}) {
+	By("creating related resources")
+	for i := range input.RelatedResources {
+		obj := input.RelatedResources[i]
+		By(fmt.Sprintf("creating a/an %s resource", obj.GetObjectKind().GroupVersionKind()))
+		Eventually(func() error {
+			return input.Creator.Create(ctx, obj)
+		}, intervals...).Should(Succeed())
 	}
 }
