--- test/framework/types.go
@@ -1,27 +1,0 @@
-/*
-Copyright 2019 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package framework
-
-import (
-	"reflect"
-)
-
-// TypeToKind returns the Kind without the package prefix. Pass in a pointer to a struct
-// This will panic if used incorrectly.
-func TypeToKind(i interface{}) string {
-	return reflect.ValueOf(i).Elem().Type().Name()
-}
